Revert "Refactor Approve Fn and Add Jest Test  (#300)"
This reverts commit b341599c8f6807b8f1b8ff4cb0a7cf409f5eae92.
